# gitrank/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GitRank Gateway API
  version: 0.2.0
  description: |
    Public and authenticated HTTP surface served by `api-gateway`.

    Compatibility policy:
    - all externally supported routes are versioned under `/v1`
    - additive changes may add optional fields without a version bump
    - breaking changes require a new versioned route family and a documented migration window

    Conventions:
    - authenticated browser clients use the `gitrank_session` cookie
    - state-changing routes additionally require `X-CSRF-Token`
    - public profile responses are cacheable for a short window
    - authenticated profile and sync responses are returned with `Cache-Control: private, no-store`
    - rate limiting currently exposes `Retry-After` on `429 Too Many Requests`

    Pagination and filtering:
    - current `/v1` routes do not expose paginated collections
    - future collection routes should use `page[size]`, `page[after]`, `page[before]`, `sort`, and `filter[...]`
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/me/account/export:
    get:
      tags: [Account]
      summary: Export authenticated account-owned GitRank data without token secrets
      description: |
        Returns the authenticated user's profile, score history, badges, repository visibility,
        linked GitHub account metadata, sanitized session metadata, and sanitized audit events.
        Raw session tokens, token hashes, CSRF hashes, encrypted GitHub tokens, and refresh tokens
        are intentionally excluded.
      operationId: exportMyAccountData
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Account data export generated
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDataExportResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/me/profile:
    get:
      tags: [Profiles]
      summary: Return the authenticated user's profile read model
      operationId: getMyProfile
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated profile snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateProfileResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Profiles]
      summary: Update authenticated profile privacy controls
      operationId: updateMyProfilePrivacy
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfilePrivacyRequest"
      responses:
        "200":
          description: Updated authenticated profile snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateProfileResponse"
        "400":
          description: Invalid privacy update
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid or missing CSRF token
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/quests:
    get:
      tags: [Profiles]
      summary: Return authenticated quest recommendations derived from profile score evidence
      description: |
        Quest recommendations are materialized from the latest profile snapshot and verified score history.
        Completed quests write progress, completion, reward grant, badge, score-event, and audit evidence.
      operationId: getMyQuests
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current authenticated quest board
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuestsResponse"
        "401":
          description: Authentication required
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/pr/{owner}/{repo}/{number}/report/materialize:
    post:
      tags: [Profiles]
      summary: Persist an idempotent PR battle report snapshot from current evidence
      operationId: materializePullRequestReport
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: number
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "202":
          description: Pull request report snapshot materialized
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequestReportMaterializationResponse"
        "404":
          description: Pull request report not found or not public
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/leaderboard:
    get:
      tags: [Profiles]
      summary: Return the public GitRank leaderboard
      description: |
        The leaderboard is built from public, leaderboard-enabled profile snapshots materialized into
        weekly leaderboard season snapshots and rank movement events.
        Placement is directional and formula-versioned, not an absolute measure of engineering ability.
      operationId: getLeaderboard
      responses:
        "200":
          description: Public leaderboard snapshot
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Cache-Control:
              $ref: "#/components/headers/PublicProfileCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/leaderboard/materialize:
    post:
      tags: [Profiles]
      summary: Materialize the current leaderboard season
      description: |
        Rebuilds the current weekly leaderboard season snapshots and rank movement events from public profile snapshots.
        Intended for scheduler-worker and repair flows; public reads can still serve the latest materialized season.
      operationId: materializeLeaderboardSeason
      responses:
        "202":
          description: Leaderboard season materialized
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardMaterializationResponse"
        "429":
          description: Rate limit exceeded
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  headers:
    XRequestID:
      description: Request correlation ID injected at the gateway edge.
      schema:
        type: string
    RetryAfter:
      description: Seconds to wait before retrying a rate-limited request.
      schema:
        type: integer
        minimum: 1
    PublicProfileCacheControl:
      description: Public cache policy applied to public profile responses.
      schema:
        type: string
        example: public, max-age=60, stale-while-revalidate=300
    PrivateNoStoreCacheControl:
      description: Private no-store cache policy applied to authenticated and sync responses.
      schema:
        type: string
        example: private, no-store
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gitrank_session
      description: "Session cookie issued by `auth-service`."
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Double-submit CSRF token derived from the current session.
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    AccountDataExportResponse:
      type: object
      required: [export_version, generated_at, user, profile]
      properties:
        export_version:
          type: string
          example: account-export/v1
        generated_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/AccountExportUser"
        github_accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountExportGitHubAccount"
        profile:
          $ref: "#/components/schemas/PrivateProfileResponse"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/AccountExportSession"
        audit_events:
          type: array
          items:
            $ref: "#/components/schemas/AccountExportAuditEvent"
        redactions:
          type: array
          items:
            type: string
          description: Human-readable notes for omitted or redacted secret-bearing fields.
    AccountExportUser:
      type: object
      required: [user_id, public_handle, display_name, status, profile_visibility, created_at, updated_at]
      properties:
        user_id:
          type: string
          format: uuid
        public_handle:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
        bio:
          type: string
        status:
          type: string
        profile_visibility:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountExportGitHubAccount:
      type: object
      required: [github_account_id, github_user_id, login, access_mode, installation_count, link_status, linked_at, created_at, updated_at]
      properties:
        github_account_id:
          type: string
          format: uuid
        github_user_id:
          type: integer
          format: int64
        login:
          type: string
        display_name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
        user_type:
          type: string
        access_mode:
          type: string
        oauth_scopes:
          type: array
          items:
            type: string
        installation_count:
          type: integer
        link_status:
          type: string
        linked_at:
          type: string
          format: date-time
        unlinked_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AccountExportSession:
      type: object
      required: [session_id, github_account_id, github_authorization_status, created_at, last_seen_at, last_refreshed_at, rotated_at, expires_at, idle_expires_at]
      properties:
        session_id:
          type: string
          format: uuid
        github_account_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
        request_ip:
          type: string
        user_agent:
          type: string
        github_authorization_status:
          type: string
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        last_refreshed_at:
          type: string
          format: date-time
        rotated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        idle_expires_at:
          type: string
          format: date-time
        invalidated_at:
          type: string
          format: date-time
        invalidated_reason:
          type: string
    AccountExportAuditEvent:
      type: object
      required: [id, actor_type, action, target_type, created_at]
      properties:
        id:
          type: string
          format: uuid
        actor_type:
          type: string
        actor_id:
          type: string
        action:
          type: string
        target_type:
          type: string
        target_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    PublicProfileSummary:
      type: object
      required: [handle, display_name, total_xp, strength_summary, badges_earned, merged_pull_requests, updated_at]
      properties:
        handle:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
        bio:
          type: string
        total_xp:
          type: integer
        strength_summary:
          type: string
          description: Directional, evidence-backed summary of current contribution strengths.
        top_skills:
          type: array
          items:
            type: string
        badges_earned:
          type: integer
        merged_pull_requests:
          type: integer
        updated_at:
          type: string
          format: date-time
...
    ProfileLevelView:
      type: object
      required: [label, current_level, current_xp, next_level_xp, rank_tier]
      properties:
        label:
          type: string
        current_level:
          type: integer
        current_xp:
          type: integer
        next_level_xp:
          type: integer
        rank_tier:
          type: string
...
    LeaderboardResponse:
      type: object
      required: [entries, window, generated_at]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntryView"
        window:
          $ref: "#/components/schemas/ProfileTimeWindow"
        generated_at:
          type: string
          format: date-time
        season_key:
          type: string
        season_snapshot_version:
          type: string
        scoring_version:
          type: string
    LeaderboardMaterializationResponse:
      type: object
      required: [status, season_key, season_snapshot_version, entry_count, window, source_watermark, generated_at]
      properties:
        status:
          type: string
          enum: [materialized]
        season_key:
          type: string
        season_snapshot_version:
          type: string
        scoring_version:
          type: string
        entry_count:
          type: integer
          minimum: 0
        window:
          $ref: "#/components/schemas/ProfileTimeWindow"
        source_watermark:
          type: string
          format: date-time
        generated_at:
          type: string
          format: date-time
...
    UserQuestsResponse:
      type: object
      required: [quests, generated_at, staleness]
      properties:
        quests:
          type: array
          items:
            $ref: "#/components/schemas/QuestView"
        generated_at:
          type: string
          format: date-time
        staleness:
          $ref: "#/components/schemas/ProfileStaleness"
...
    PullRequestReportMaterializationResponse:
      type: object
      required:
        - status
        - repository
        - number
        - pull_request_id
        - report_snapshot_id
        - report_version
        - is_stale
        - source_updated_at
        - generated_at
      properties:
        status:
          type: string
        repository:
          type: string
        number:
          type: integer
          minimum: 1
        pull_request_id:
          type: string
          format: uuid
        report_snapshot_id:
          type: string
          format: uuid
        report_version:
          type: string
        score_event_id:
          type: string
          format: uuid
        analysis_id:
          type: string
          format: uuid
        score_version:
          type: string
        analysis_version:
          type: string
        evidence_status:
          type: string
        missing_evidence:
          type: array
          items:
            type: string
        is_stale:
          type: boolean
        source_updated_at:
          type: string
          format: date-time
        generated_at:
          type: string
          format: date-time
...
    PrivateProfileResponse:
      type: object
      required: [summary, level, timeline, privacy, share_card, staleness]
      properties:
        summary:
          $ref: "#/components/schemas/PublicProfileSummary"
        top_skill_areas:
          type: array
          items:
            $ref: "#/components/schemas/SkillAreaView"
        top_repositories:
          type: array
          items:
            $ref: "#/components/schemas/TopRepositoryView"
        level:
          $ref: "#/components/schemas/ProfileLevelView"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/BadgeView"
        timeline:
          $ref: "#/components/schemas/ProfileTimeline"
        score_history:
          type: array
          items:
            $ref: "#/components/schemas/ScoreHistoryEntry"
        recent_pr_reports:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestReportResponse"
        privacy:
          $ref: "#/components/schemas/ProfilePrivacySettings"
        repository_visibility:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryVisibilityView"
        share_card:
          $ref: "#/components/schemas/ShareableProfileCard"
        staleness:
          $ref: "#/components/schemas/ProfileStaleness"
    UpdateProfilePrivacyRequest:
      type: object
      properties:
        public_profile_enabled:
          type: boolean
        show_exact_prs:
          type: boolean
        show_ai_summaries:
          type: boolean
        show_leaderboard_participation:
          type: boolean
        reduced_gamification:
          type: boolean
...